+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/lib/argument/filter/ArgumentFilter.java
@@ -1,10 +1,10 @@
 package com.nisovin.shopkeepers.commands.lib.argument.filter;
 
-import java.util.function.Predicate;
-
 import org.checkerframework.checker.nullness.qual.Nullable;
 
+import com.nisovin.shopkeepers.commands.lib.CommandInput;
 import com.nisovin.shopkeepers.commands.lib.argument.CommandArgument;
+import com.nisovin.shopkeepers.commands.lib.context.CommandContextView;
 import com.nisovin.shopkeepers.text.Text;
 
 /**
@@ -13,11 +13,32 @@ import com.nisovin.shopkeepers.text.Text;
  * @param <T>
  *            the type of the filtered parsed arguments
  */
-public abstract class ArgumentFilter<T> implements Predicate<T> {
+public abstract class ArgumentFilter<T> {
+
+	/**
+	 * Evaluates this filter on the given argument.
+	 * 
+	 * @param input
+	 *            the command input, not <code>null</code>
+	 * @param context
+	 *            the context which stores the parsed argument values, not <code>null</code>
+	 * @param value
+	 *            the parsed argument value
+	 * @return {@code true} if the argument value is accepted by this filter
+	 */
+	public abstract boolean test(
+			CommandInput input,
+			CommandContextView context,
+			T value
+	);
 
 	private static final ArgumentFilter<@Nullable Object> ACCEPT_ANY = new ArgumentFilter<@Nullable Object>() {
 		@Override
-		public boolean test(@Nullable Object value) {
+		public boolean test(
+				CommandInput input,
+				CommandContextView context,
+				@Nullable Object value
+		) {
 			return true;
 		}
 	};